Error:  [06:19:37] nodejs-1           | /code/frontend/src/js/controllers/buddhismLeaderboardCtrl.js nodejs-1           |   158:44  error  'challengeId' is defined but never used  no-unused-vars nodejs-1           | nodejs-1           | ✖ 1 problem (1 error, 0 warnings) nodejs-1           | nodejs-1           | Please remove lint errors to connect the server nodejs-1           | [06:19:37] 'lint' errored after 2.39 s nodejs-1           | [06:19:37] ESLintError in plugin "gulp-eslint" nodejs-1           | Message: nodejs-1           |     Failed with 1 error nodejs-1           | Details: nodejs-1           |     domainEmitter: [object Object] nodejs-1           |     domainThrown: false
solution:
The original code was likely designed to generate different chart data based on the challenge ID, but in the current implementation, the parameter wasn't actually being used to customize the data. The function was generating random data regardless of which challenge ID was passed to it.

After this fix, your server should start without the linting error. If you do want to customize chart data based on the challenge ID in the future, you can reintroduce the parameter and use it within the function to generate challenge-specific data.